Add PictureRecorder components
