+++ example/src/Examples/API/Picture.tsx
@@ -1,6 +1,5 @@
-import type { SkPicture, SkRect } from "@shopify/react-native-skia";
+import type { SkPicture } from "@shopify/react-native-skia";
 import {
-  createDrawing,
   rect,
   Skia,
   Canvas,
@@ -8,8 +7,8 @@ import {
   Group,
   Rect,
   Circle,
+  PictureRecorder,
 } from "@shopify/react-native-skia";
-import type { ReactNode } from "react";
 import React, { useMemo, useState } from "react";
 import { StyleSheet } from "react-native";
 
@@ -32,13 +31,10 @@ export const PictureExample = () => {
 
   return (
     <Canvas style={styles.container}>
-      <RecordPicture
-        onRecord={(pic) => setPicture(pic)}
-        bounds={rect(0, 0, 100, 100)}
-      >
+      <PictureRecorder onRecord={setPicture} bounds={rect(0, 0, 100, 100)}>
         <Rect x={0} y={0} width={100} height={100} color="pink" />
         <Circle cx={50} cy={50} r={50} color="orange" />
-      </RecordPicture>
+      </PictureRecorder>
       <Picture picture={picture} />
       <Group transform={[{ translateX: 200 }]}>
         {copyOfPicture && <Picture picture={copyOfPicture} />}
@@ -52,31 +48,3 @@ const styles = StyleSheet.create({
     flex: 1,
   },
 });
-
-interface RecordPictureProps {
-  onRecord: (picture: SkPicture) => void;
-  bounds: SkRect;
-  children?: ReactNode | ReactNode[];
-}
-
-const RecordPicture = ({ onRecord, bounds, ...props }: RecordPictureProps) => {
-  const [picture, setPicture] = useState<null | SkPicture>(null);
-  const onDraw = useMemo(
-    () =>
-      createDrawing<RecordPictureProps>((ctx, {}, node) => {
-        if (picture === null) {
-          const recorder = Skia.PictureRecorder();
-          const canvas = recorder.beginRecording(bounds);
-          node.visit({
-            ...ctx,
-            canvas,
-          });
-          const pic = recorder.finishRecordingAsPicture();
-          setPicture(pic);
-          onRecord(pic);
-        }
-      }),
-    [bounds, onRecord, picture]
-  );
-  return <skDrawing onDraw={onDraw} skipProcessing {...props} />;
-};
