+++ package/src/renderer/components/index.ts
@@ -9,6 +9,7 @@ export * from "./imageFilters";
 export * from "./pathEffects";
 export * from "../processors";
 export * from "./Picture";
+export * from "./PictureRecorder";
 
 export * from "./Group";
 export * from "./Mask";
